update struct folder
