+++ .env.development
@@ -1,0 +1,1 @@
+REACT_APP_BASE_URL=https://tiktok.fullstack.edu.vn/api
