+++ src/App.js
@@ -1,7 +1,7 @@
 import { Fragment } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { publicRoutes } from '~/routes';
-import { DefaultLayout } from '~/layouts';
+import { publicRoutes } from '~/routes/routes';
+import DefaultLayout from '~/layouts';
 function App() {
    return (
       <Router>
