+++ src/components/GlobalStyles/GlobalStyles.scss
@@ -9,6 +9,7 @@
    --white: #f8f8f8;
    --black: #000;
    --text-color: #333;
+   --light-gray: rgb(243, 243, 243);
 
    --default-layout-header-height: 60px;
    --default-layout-width: 1200px;
