+++ src/components/GlobalStyles/index.js
@@ -1,5 +1,1 @@
-import './GlobalStyles.scss';
-function GlobalStyles({ children }) {
-  return children;
-}
-export default GlobalStyles;
+export { default } from './GlobalStyles';
