+++ src/layouts/DefaultLayout/DefaultLayout.js
@@ -1,0 +1,24 @@
+import PropTypes from 'prop-types';
+import classNames from 'classnames/bind';
+import Header from '~/layouts/components/Header';
+import Footer from '~/layouts/components/Footer';
+
+import styles from './Default.module.scss';
+const cx = classNames.bind(styles);
+
+function DefaultLayout({ children }) {
+   return (
+      <div className={cx('wrapper')}>
+         <Header />
+         <div className={cx('container')}>
+            <div className={cx('content')}>{children}</div>
+         </div>
+         <Footer />
+      </div>
+   );
+}
+DefaultLayout.propTypes = {
+   children: PropTypes.node.isRequired,
+};
+
+export default DefaultLayout;
