+++ src/layouts/DefaultLayout/index.js
@@ -1,20 +1,1 @@
-import classNames from 'classnames/bind';
-import Header from '~/layouts/components/Header';
-import Footer from '~/layouts/components/Footer';
-
-import styles from './Default.module.scss';
-const cx = classNames.bind(styles);
-
-function DefaultLayout({ children }) {
-   return (
-      <div className={cx('wrapper')}>
-         <Header />
-         <div className={cx('container')}>
-            <div className={cx('content')}>{children}</div>
-         </div>
-         <Footer />
-      </div>
-   );
-}
-
-export default DefaultLayout;
+export { default } from './DefaultLayout';
