+++ src/layouts/components/FeaturedProduct/FeaturedProduct/FeaturedProductItem.module.scss
@@ -1,0 +1,2 @@
+.wrapper {
+}
