+++ src/layouts/components/FeaturedProduct/FeaturedProduct/index.js
@@ -1,0 +1,1 @@
+export { default } from './FeaturedProductItem';
