+++ src/layouts/components/SearchTag/SearchTag.module.scss
@@ -1,0 +1,7 @@
+.wrapper {
+   width: 100%;
+   height: 100%;
+   display: flex;
+   align-items: center;
+   justify-content: right;
+}
