+++ src/layouts/components/Swiper/Slide/index.js
@@ -1,13 +1,1 @@
-import classNames from 'classnames/bind';
-import styles from './Slide.module.scss';
-const cx = classNames.bind(styles);
-
-function Slide(props) {
-   return (
-      <div className={cx('wrapper')}>
-         <img src={props.item.imgSource} alt={props.alt} />
-      </div>
-   );
-}
-
-export default Slide;
+export { default } from './Slide';
