+++ src/layouts/index.js
@@ -1,1 +1,1 @@
-export { default as DefaultLayout } from './DefaultLayout';
+export { default } from './DefaultLayout';
